pull out qc chip into own component
